backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pandas==2.2.2
chardet==5.2.0
stripe==9.3.0
python-dotenv==1.0.1
python-multipart==0.0.9
scipy==1.13.0
